+++ Model/QliroOrder/Builder/CreateRequestBuilder.php
@@ -21,7 +21,6 @@ use Qliro\QliroOne\Model\Config;
 use Qliro\QliroOne\Model\Logger\Manager;
 use Qliro\QliroOne\Model\Management\CountrySelect;
 use Qliro\QliroOne\Service\Callback\UrlBuilder as CallbackUrlBuilder;
-use Magento\Store\Model\Information;
 
 /**
  * QliroOne Order create request builder class
@@ -94,11 +93,6 @@ class CreateRequestBuilder
     private $shippingMethodsBuilder;
 
     /**
-     * @var \Magento\Store\Model\Information
-     */
-    private $information;
-
-    /**
      * @var \Magento\Framework\Event\ManagerInterface
      */
     private $eventManager;
@@ -133,7 +127,6 @@ class CreateRequestBuilder
      * @param \Qliro\QliroOne\Api\GeoIpResolverInterface $geoIpResolver
      * @param \Qliro\QliroOne\Service\Callback\UrlBuilder $callbackUrlBuilder
      * @param \Qliro\QliroOne\Model\QliroOrder\Builder\ShippingMethodsBuilder $shippingMethodsBuilder
-     * @param \Magento\Store\Model\Information $information
      * @param \Magento\Framework\Event\ManagerInterface $eventManager
      * @param \Qliro\QliroOne\Model\Management\CountrySelect $countrySelect
      * @param Manager $logManager
@@ -152,7 +145,6 @@ class CreateRequestBuilder
         CallbackUrlBuilder $callbackUrlBuilder,
         ShippingMethodsBuilder $shippingMethodsBuilder,
         ShippingConfigBuilder $shippingConfigBuilder,
-        Information $information,
         ManagerInterface $eventManager,
         CountrySelect $countrySelectManagement,
         Manager $logManager
@@ -169,7 +161,6 @@ class CreateRequestBuilder
         $this->geoIpResolver = $geoIpResolver;
         $this->callbackUrlBuilder = $callbackUrlBuilder;
         $this->shippingMethodsBuilder = $shippingMethodsBuilder;
-        $this->information = $information;
         $this->eventManager = $eventManager;
         $this->shippingConfigBuilder = $shippingConfigBuilder;
         $this->countrySelectManagement = $countrySelectManagement;
@@ -210,36 +201,17 @@ class CreateRequestBuilder
 
         $this->logManager->debug('Starting to set order items to request object');
         $createRequest->setOrderItems($orderItems);
-        $presetAddress = $this->qliroConfig->presetAddress();
+        /*
+         * The preset shipping address lives in ShippingMethodsBuilder, at the one place that
+         * rates. The placeholder has to be there for every rating in the request, not only for
+         * this one, and it must never stay on the quote: what is left on the quote is what the
+         * order is placed with, and the store name reached the buyer's order that way.
+         */
         $shippingAddress = $this->quote->getShippingAddress();
-        if ($presetAddress && empty($shippingAddress->getPostcode())) {
-            $this->logManager->debug('Starting to set fake address as don\'t have real one yet');
-            /* set a fake address since we don't have the real one yet */
-            $storeInfo = $this->information->getStoreInformationObject($this->quote->getStore());
-            if (!empty($storeInfo)) {
-                $shippingAddress->addData([
-                    'company' => $storeInfo->getData('name'),
-                    'telephone' => $storeInfo->getData('phone'),
-                    'street' => sprintf(
-                        "%s\n%s",
-                        $storeInfo->getData('street_line1'),
-                        $storeInfo->getData('street_line2')
-                    ),
-                    'city' => $storeInfo->getData('city'),
-                    'postcode' => str_replace(' ', '', (string)$storeInfo->getData('postcode')),
-                    'region_id' => $storeInfo->getData('region_id'),
-                    'country_id' => $storeInfo->getData('country_id'),
-                    'region' => $storeInfo->getData('region'),
-                ]);
-            }
-        }
         $shippingAddress->setCollectShippingRates(true)->collectShippingRates()->save();
         $this->logManager->debug('Starting to get shipping methods for quote: ' . $this->quote->getId());
         $shippingMethods = $this->shippingMethodsBuilder->setQuote($this->quote)->create();
         $availableShippingMethods = $shippingMethods->getAvailableShippingMethods();
-        if (!empty($storeInfo)) {
-            $shippingAddress->clearInstance()->save();
-        }
         $createRequest->setAvailableShippingMethods($availableShippingMethods);
 
         $shippingConfig = $this->shippingConfigBuilder->setQuote($this->quote)->create();